Split CLI and MCP sources into documented module files
CLI crate:
- src/args.rs — Args struct, Command enum, all arg enums + From impls
- src/main.rs — main(), parse_files_list(), run_scrub_check(), tests

MCP crate:
- src/tools.rs — CargoContextServer struct, tool arg types,
  #[tool_router] impl with 4 tool handlers
- src/server.rs — #[tool_handler] ServerHandler impl (resources,
  prompts, get_info)
- src/main.rs — tokio main(), tracing init, stdio transport

Closes #9
